% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{testfun}
\alias{testfun}
\title{2D test function for sequential optimization}
\usage{
testfun(X)
}
\arguments{
\item{X}{A matrix}
}
\value{
A vector of function outputs.
}
\description{
The function is defined on the unit interval [0,1]^2.
}
\examples{
X <- matrix(c(0.2,0.3,0.1,0.2), ncol=2)
testfun(X)

}
